Responsive Profile and Login-Resgitration Update
I make the profile page responsive.

I also prepare the login page and registration page for connecting with authentication context so that when we click the Sign In and Sign Up buttons, thing will go to the right place as well as updated on user level.

I also update the login page for navigation if it was directed from the movie page when the guests click on movie.